+++ R/dist_multivariate_normal.R
@@ -10,6 +10,7 @@
 #'
 #' @examples
 #' dist <- dist_multivariate_normal(mu = list(c(1,2)), sigma = list(matrix(c(4,2,2,3), ncol=2)))
+#' dimnames(dist) <- c("x", "y")
 #' dist
 #'
 #' @examplesIf requireNamespace("mvtnorm", quietly = TRUE)
@@ -18,12 +19,13 @@
 #' support(dist)
 #' generate(dist, 10)
 #'
-#' density(dist, c(2, 1))
-#' density(dist, c(2, 1), log = TRUE)
+#' density(dist, cbind(2, 1))
+#' density(dist, cbind(2, 1), log = TRUE)
 #'
 #' cdf(dist, 4)
 #'
 #' quantile(dist, 0.7)
+#' quantile(dist, 0.7, type = "marginal")
 #'
 #' @export
 dist_multivariate_normal <- function(mu = 0, sigma = diag(1)){
@@ -54,26 +56,28 @@ log_density.dist_mvnorm <- function(x, at, ..., na.rm = FALSE){
 }
 
 #' @export
-quantile.dist_mvnorm <- function(x, p, type = c("univariate", "equicoordinate"),
+quantile.dist_mvnorm <- function(x, p, type = c("marginal", "equicoordinate"),
                                  ..., na.rm = FALSE){
   type <- match.arg(type)
-  if (type == "univariate") {
-    matrix(
-      stats::qnorm(p, mean = rep(x[["mu"]], each = length(p)),
-                   sd = rep(diag(sqrt(x[["sigma"]])), each = length(p)), ...),
-      nrow = length(p)
-    )
+  q <- if (type == "marginal") {
+    stats::qnorm(p, mean = rep(x[["mu"]], each = length(p)),
+                 sd = rep(sqrt(diag(x[["sigma"]])), each = length(p)), ...)
   } else {
     require_package("mvtnorm")
-    mvtnorm::qmvnorm(p, mean = x[["mu"]], sigma = x[["sigma"]], ...)$quantile
+    vapply(p, function(p, ...) {
+      if (p == 0) return(-Inf) else if (p == 1) return(Inf)
+      mvtnorm::qmvnorm(p, ...)$quantile
+    }, numeric(1L), mean = x[["mu"]], sigma = x[["sigma"]], ...)
   }
+
+  matrix(q, nrow = length(p), ncol = dim(x))
 }
 
 #' @export
 cdf.dist_mvnorm <- function(x, q, ..., na.rm = FALSE){
   if(is.list(q)) return(vapply(q, cdf, numeric(1L), x = x, ...))
   require_package("mvtnorm")
-  mvtnorm::pmvnorm(as.numeric(q), mean = x[["mu"]], sigma = x[["sigma"]], ...)[1]
+  mvtnorm::pmvnorm(upper = as.numeric(q), mean = x[["mu"]], sigma = x[["sigma"]], ...)[1]
 }
 
 #' @export